requirements.txt
django
django-debug-toolbar
requests-oauthlib
